mnist/requirements.txt
tensorflow>=2.18.0,<3.0.0
numpy>=2.0.2
matplotlib>=3.9.2
seaborn>=0.13.2
scikit-learn>=1.5.2
PyQt6>=6.7.1
Pillow>=11.0.0